Transfer unique values from backend in binary
